sports_data_service/requirements.txt
# Core Django
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0

# Database
psycopg2-binary>=2.9.0,<3.0.0
psycopg2-pool>=1.1.0,<2.0.0

# JWT Authentication
PyJWT>=2.8.0,<3.0.0

# System Monitoring
psutil>=5.9.0,<6.0.0

# HTTP Requests
requests>=2.31.0,<3.0.0

# Caching
redis>=4.6.0,<5.0.0
django-redis>=5.3.0,<6.0.0

# Environment Variables
python-dotenv>=1.0.0,<2.0.0
python-decouple>=3.8,<4.0.0

# Logging
structlog>=23.1.0,<24.0.0

# Development Tools
django-debug-toolbar>=4.2.0,<5.0.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-django>=4.5.0,<5.0.0

# Code Quality
black>=23.7.0,<24.0.0
flake8>=6.0.0,<7.0.0
isort>=5.12.0,<6.0.0

# Documentation
Sphinx>=7.1.0,<8.0.0
sphinx-rtd-theme>=1.3.0,<2.0.0
